Installation of GBP from stackforge
Sets up GBP installation as a part of devstack installation.

This includes cloning of the server and client repos, and running
the DB migration script to create the GBP tables in the Neutron DB.

Since GBP patches are still in review, local.conf overrides stackforge
repo references with links to the gerrit patches.

Conflicts:
	stackrc