Install futures==2.2.0 manually
A recent change to taskflow (https://review.openstack.org/#/c/212652/)
pinned the dependency of futures to <=2.2.0 and was released as
taskflow==0.6.2 a couple days ago, which installed as a dependency of
Cinder (cinder requires taskflow<0.7).

However, futures is installed as a dependency of python-swiftclient,
which does not pin futures, so futures==3.0.3 is installed.

When a cinder command is run, the dependency check of taskflow fails
since the newer version of futures is installed.

It seems infeasible to update the requirements everywhere futures is
required (its more than *just* python-swiftclient and I'm not certain we
can rely on python-swiftclient being installed first), so instead,
install futures==2.2.0 before attempting to install other packages.